Add `persistent_workers` argument to dataloaders (slightly experimental)
Ideally, this should noticeably speed up the first batch of every epoch
after the first one by getting rid of the overhead for re-creating the
worker processes. It seems that this affects the batches that are being
produced, but the unit test suggests it is still reproducible. Let's see
how well this works in practice...